requirements.txt
scikit-learn

numpy==1.22.3
pandas==1.4.2
scipy==1.8.0
matplotlib==3.5.2
opencv-python==4.5.5.64
Pillow==9.1.0

# If you have an X86 CPU architecture use tensorflow==2.8.0 instead
tensorflow-macos==2.8.0
tensorflow-metal

keras

gitpython==3.1.27
selenium~=3.141.0
requests~=2.27.1

tqdm~=4.64.0

notebook
jupyterlab